Move CreateOrderOpts into shared types
